moves the badge, button and checkbox component in their own directory with the tests
